Move API calls of SQS/SNS from models to controller
