+++ app/models/barbeque/job_queue.rb
@@ -18,44 +18,4 @@ class Barbeque::JobQueue < Barbeque::ApplicationRecord
   def sqs_queue_name
     SQS_NAME_PREFIX + name
   end
-
-  def update_policy!
-    client = Aws::SQS::Client.new
-    attrs = client.get_queue_attributes(
-      queue_url: queue_url,
-      attribute_names: ['QueueArn'],
-    )
-    queue_arn = attrs.attributes['QueueArn']
-    topic_arns = sns_subscriptions.map(&:topic_arn)
-    if topic_arns.present?
-      policy = generate_policy(queue_arn: queue_arn, topic_arns: topic_arns)
-    end
-
-    client.set_queue_attributes(
-      queue_url: queue_url,
-      attributes: { 'Policy' => policy },
-    )
-  end
-
-  private
-
-  # @paaram [String] queue_arn
-  # @param [Array<String>] topic_arns
-  # @return [String] JSON formatted policy
-  def generate_policy(queue_arn:, topic_arns:)
-    {
-      'Version' => '2012-10-17',
-      'Statement' => [
-        'Effect' => 'Allow',
-        'Principal' => '*',
-        'Action' => 'sqs:SendMessage',
-        'Resource' => queue_arn,
-        'Condition' => {
-          'ArnEquals' => {
-            'aws:SourceArn' => topic_arns,
-          }
-        }
-      ]
-    }.to_json
-  end
 end
